Reset culture to invariant
Running the server on a Czech system (and most likely also other non-English cultures) throws an exception in `Extensions.SetValueEx`. This is most likely caused by `Convert.ChangeType` trying to convert a floating-point value, however the affected cultures use a comma as the default decimal separator, resulting in the failure.

This change sets the culture of the application and all new threads to one that behaves consistently on all platforms.